requirements.txt
asgiref==3.7.2
boto3==1.34.32
botocore==1.34.32
certifi==2023.11.17
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-jazzmin==2.6.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
packaging==23.2
pillow==10.2.0
platformdirs==4.1.0
psycopg2==2.9.6
PyJWT==2.8.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
qrcode==7.4.2
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
stripe==7.13.0
typing_extensions==4.9.0
urllib3==2.0.7
virtualenv==20.25.0
